Add setQuestionsTv function to handle TV show trivia questions
